requirements.txt
# Core ML
xgboost>=2.0
scikit-learn>=1.4
pandas>=2.1
numpy>=1.26
pyarrow>=14.0

# Data / HTTP
requests>=2.31
python-dotenv>=1.0

# API server
fastapi>=0.111
uvicorn[standard]>=0.29
pydantic>=2.0

# Database
sqlalchemy>=2.0
psycopg2-binary>=2.9

# GCP
google-cloud-storage>=2.16

# Email
sendgrid>=6.11
